+++ crates/preset_env_base/src/lib.rs
@@ -1,3 +1,11 @@
+use anyhow::Error;
+use serde::Deserialize;
+use st_map::StaticMap;
+
+pub use self::version::Version;
+
+mod version;
+
 /// A map without allocation.
 #[derive(Debug, Default, Deserialize, Clone, Copy, StaticMap)]
 #[serde(deny_unknown_fields)]
@@ -36,4 +44,55 @@ pub struct BrowserData<T: Default> {
     pub opera_mobile: T,
     #[serde(default)]
     pub rhino: T,
+}
+
+pub type Versions = BrowserData<Option<Version>>;
+
+impl BrowserData<Option<Version>> {
+    pub(crate) fn is_any_target(&self) -> bool {
+        self.iter().all(|(_, v)| v.is_none())
+    }
+
+    pub(crate) fn parse_versions(distribs: Vec<browserslist::Distrib>) -> Result<Self, Error> {
+        fn remap(key: &str) -> &str {
+            match key {
+                "and_chr" => "chrome",
+                "and_ff" => "firefox",
+                "ie_mob" => "ie",
+                "ios_saf" => "ios",
+                "op_mob" => "opera",
+                _ => key,
+            }
+        }
+
+        let mut data: Versions = BrowserData::default();
+        for dist in distribs {
+            let browser = dist.name();
+            let browser = remap(browser);
+            let version = dist.version();
+            match &*browser {
+                "and_qq" | "and_uc" | "baidu" | "bb" | "kaios" | "op_mini" => continue,
+
+                _ => {}
+            }
+
+            let version = version
+                .split_once('-')
+                .map(|(version, _)| version)
+                .unwrap_or(version)
+                .parse()
+                .unwrap();
+
+            // lowest version
+            if data[&browser].map(|v| v > version).unwrap_or(true) {
+                for (k, v) in data.iter_mut() {
+                    if browser == k {
+                        *v = Some(version);
+                    }
+                }
+            }
+        }
+
+        Ok(data)
+    }
 }
