+++ crates/swc_ecma_preset_env/Cargo.toml
@@ -11,8 +11,6 @@ version = "0.93.0"
 
 [dependencies]
 ahash = "0.7.4"
-anyhow = "1"
-browserslist-rs = "=0.8.0"
 dashmap = "4.0.2"
 indexmap = "1.6.2"
 once_cell = "1.9.0"
